feat: run hooks through a shell (fixes #19)
This commit introduces a shell command option on which the actual hook commands will be run.

It works by replacing a `{}` placeholder. The default, for example, is `sh -c '{}'`, which should support most POSIX features you need.

If you need a specific shell feature, just set it to whatever you want! If for some reason you want the old behaviour (maybe you are crazy enough to use flavours on Windows?), just set it to `{}` and hooks will be ran "directly"